refactor seq logic create useStepActions hook remove logic from container component
