+++ tsequencer-frontend/src/components/StepSequencerContainer/components/StepSequencer/index.tsx
@@ -2,6 +2,7 @@ import StepSeqButton from "./components/StepSequencerButton";
 
 import type { windowSize } from "@/types";
 import type { Sequencer } from "../../types";
+import LoadingSpinner from "@/components/UtilityComponents/LoadingSpinner";
 
 const StepSequencer = ({
   seq,
@@ -13,7 +14,7 @@ const StepSequencer = ({
   windowSize: windowSize;
 }) => {
   // TODO: Proper Loading state
-  if (!seq || !seq[0]) return <>Loading...</>;
+  if (!seq || !seq[0]) return <LoadingSpinner />;
   return (
     <div className="seq-container grid gap-4 grid-cols-4 grid-rows-4">
       {seq.map((b, i) => {
